+++ index.js
@@ -15,7 +15,30 @@ function isEmail(email) {
   } else {
     console.log("Non- Validated Email");
   }
-  return bool
+  return bool;
 }
 
-module.exports = { add, minus, isEmail };
+function passwordStrength(password) {
+  const strength = {
+    hasUpperCase: /[A-Z]/.test(password),
+    hasLowerCase: /[a-z]/.test(password),
+    hasSpecialCharacters: /[~!@#$%^&*()_]/.test(password),
+    hasNumeric: /\d/.test(password),
+    length: password.length > 8 && password.length <= 30,
+  };
+
+  // console.log(strength);
+  let score = 0;
+  if (strength.hasUpperCase) score++;
+  if (strength.hasLowerCase) score++;
+  if (strength.hasSpecialCharacters) score++;
+  if (strength.hasNumeric) score++;
+  if (strength.length) score++;
+
+  // console.log(score)
+
+  console.log("Password strength out of 5: ", score);
+  return score
+}
+
+module.exports = { add, minus, isEmail, passwordStrength };