+++ test.js
@@ -1,5 +1,6 @@
-const { add, minus, isEmail } = require('./index')
+const { add, minus, isEmail, passwordStrength } = require('./index')
 
 console.log(add(5,6))
 console.log(minus(5,6))
 console.log(isEmail('soumya.acroy@gmail.com'))
+console.log(passwordStrength('agdajhAGH23$%'))